cfrgan-cuda/CFR-GAN/requirements.txt
torch==2.2.0+cu121
torchvision==0.17.0+cu121
pillow>=7.2.0
opencv-python==4.4.0.44
pytorch3d==0.7.5
scipy>=1.7.0
tensorboard>=2.8.0
numpy>=1.19.5